requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

EbookLib==0.18
pillow==10.4.0
requests==2.32.3
selenium==4.24.0
termcolor==2.5.0
